[geom] Refactor: move Tensor functions to _functions.py, Box.rotation_matrix
